Moved config out of root and into src
