+++ src/services/logFunctionFactory.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const config = require('../../config');
+const config = require('../config');
 const debug = require('debug');
 
 /* eslint-disable no-console */
